assetmanagement/web/requirements.txt
Django==1.11
django-daterange_filter

redis==2.10.3
django-redis==4.3.0
gunicorn==19.3.0
psycopg2==2.7.1